Move BaseContract to its own package
