+++ packages/base-contract/.npmignore
@@ -1,0 +1,5 @@
+.*
+yarn-error.log
+/scripts/
+/src/
+tsconfig.json
