+++ packages/base-contract/CHANGELOG.md
@@ -1,0 +1,5 @@
+# CHANGELOG
+
+## v0.0.1 - _TBD, 2018_
+
+    * Initial release (#TBD)
