+++ packages/base-contract/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../tsconfig",
+    "compilerOptions": {
+        "outDir": "lib"
+    },
+    "include": [
+        "./src/**/*",
+        "../../node_modules/web3-typescript-typings/index.d.ts",
+        "../../node_modules/ethers-typescript-typings/index.d.ts"
+    ]
+}
