+++ packages/base-contract/tslint.json
@@ -1,0 +1,3 @@
+{
+    "extends": ["@0xproject/tslint-config"]
+}
