+++ packages/contract_templates/contract.handlebars
@@ -4,9 +4,10 @@
  */
 // tslint:disable:no-consecutive-blank-lines
 // tslint:disable-next-line:no-unused-variable
+import { BaseContract } from '@0xproject/base-contract';
 import { TxData, TxDataPayable } from '@0xproject/types';
 import { BigNumber, classUtils, promisify } from '@0xproject/utils';
-import { BaseContract, Web3Wrapper } from '@0xproject/web3-wrapper';
+import { Web3Wrapper } from '@0xproject/web3-wrapper';
 import * as ethersContracts from 'ethers-contracts';
 import * as _ from 'lodash';
 import * as Web3 from 'web3';
